fix: very much WIP stuff, needs review
