+++ aw-sync/src/sync.rs
@@ -13,7 +13,7 @@ use std::fs;
 use std::path::{Path, PathBuf};
 
 use aw_client_rust::AwClient;
-use chrono::{DateTime, Duration, Utc};
+use chrono::{DateTime, Utc};
 
 use aw_datastore::{Datastore, DatastoreError};
 use aw_models::{Bucket, Event};
@@ -157,61 +157,6 @@ fn find_remotes_nonlocal(sync_directory: &Path, device_id: &str) -> Vec<PathBuf>
 fn create_datastore(path: &Path) -> Datastore {
     let pathstr = path.as_os_str().to_str().unwrap();
     Datastore::new(pathstr.to_string(), false)
-}
-
-// TODO: Move into tests
-fn setup_test(sync_directory: &Path) -> std::io::Result<Vec<Datastore>> {
-    let mut datastores: Vec<Datastore> = Vec::new();
-    for n in 0..2 {
-        let dspath = sync_directory.join(format!("test-remote-{}.db", n));
-        let ds_ = create_datastore(&dspath);
-        let ds = &ds_ as &dyn AccessMethod;
-
-        // Create a bucket
-        // NOTE: Created with duplicate name to make sure it still works under such conditions
-        let bucket_jsonstr = format!(
-            r#"{{
-                "id": "bucket",
-                "type": "test",
-                "hostname": "device-{}",
-                "client": "test"
-            }}"#,
-            n
-        );
-        let bucket: Bucket = serde_json::from_str(&bucket_jsonstr)?;
-        match ds.create_bucket(&bucket) {
-            Ok(()) => (),
-            Err(e) => match e {
-                DatastoreError::BucketAlreadyExists(_) => {
-                    debug!("bucket already exists, skipping");
-                }
-                e => panic!("woops! {:?}", e),
-            },
-        };
-
-        // Insert some testing events into the bucket
-        let events: Vec<Event> = (0..3)
-            .map(|i| {
-                let timestamp: DateTime<Utc> = Utc::now() + Duration::milliseconds(i * 10);
-                let event_jsonstr = format!(
-                    r#"{{
-                "timestamp": "{}",
-                "duration": 0,
-                "data": {{"test": {} }}
-            }}"#,
-                    timestamp.to_rfc3339(),
-                    i
-                );
-                serde_json::from_str(&event_jsonstr).unwrap()
-            })
-            .collect::<Vec<Event>>();
-
-        ds.insert_events(bucket.id.as_str(), events).unwrap();
-        //let new_eventcount = ds.get_event_count(bucket.id.as_str(), None, None).unwrap();
-        //info!("Eventcount: {:?} ({} new)", new_eventcount, events.len());
-        datastores.push(ds_);
-    }
-    Ok(datastores)
 }
 
 /// Returns the sync-destination bucket for a given bucket, creates it if it doesn't exist.
